[sailfish-browser] Use library for most of unit tests. JB#59794
Couple unit tests were having problems with the library so those kept
untouched, but the rest switched to using the library instead of yet
another build of the .cpp files.

Adjusted also the PKGCONFIG and similar requirements. Mostly on qmake
level, but for one code level change the MGConfItem usage on core
was made use pointer + class declaration so users of the library
don't need to themselves depend on mlite5.