[sailfish-browser] Adapt API changes to unit tests. JB#62643
Commit 7aafa69764cc8 added an extra parameter which wasn't applied
to the unit tests.

There is some failure also on tst_webview for one page state
on forwardBackwardNavigation(). The browser app side here is relatively
thin wrapper to qtmozembed so feels like the problem could be on
that side, but for starters adjusted the checks to first ensure
we're on the expected url, then ability to go back/forward and
finally the change signal amount. Earlier it was the other way around
and the reported problem was "just" amount of emitted signals.