[browser] Remove extra securityChanged signal. JB#63166
The QMozOpenGLWebPage has signal with similar name but couple parameters.
The one here doesn't seem to be emitted at all, but there's a connect()
with the name in DeclarativeWebContainer. Can't say for sure which
one gets connected but I don't make any sense of having this. Either
it's a dead thing or then it prevents the real change signal from
being handled.