[sailfish-browser] Handle F5 as reload. JB#63227
Just a simple handler for starters, doesn't bother filtering the
event from being sent to web content and doesn't do shift-f5 for hard
reload as the lower levels seem to be missing a bit there.
DeclarativeWebContainer::reload() has unused 'force' parameter,
but that's not passing the forcing to QMozOpenGLWebPage::reload() which
doesn't even have a parameter for forcing.