Make custom IconButton size overrides explicit
Filling parent is a surprising default for an item that's supposed
to be stacked on the UI. And it's also almost never used that way in the
code. Found just one place in Toolbar.qml with a placeholder parent
item that fades between two IconButtons. While height binding makes
a bit sense, all the other cases already had proper width.

So let's just make the sizes explicitly, it's not that much more
lines.