[browser] Add button for exiting fullscreen mode. JB#63258
Shown temporarily after entering fullscreen mode or tapping the screen.

Moved selection state now to qml since it's now getting a signal
when the screen is touched. What I gather there shouldn't be cases
where screen is blocked but the selection is active and web container
expected to handle the touches.

The app page side is somewhat redundant but that's the way it's been
already.