[browser] Fix some captive portal issues. JB#63259
Commit 6d45455f4 added the attribute for browser app but not the captive
portal. Thus mouse events only worked on the browser side.

Commit ff92005de added inputRegion alias to BrowserPage but not to
CaptivePortalPage.

And then bunch of things assumed having thumbnailSize / FaviconManager
which are only available, and make more sense on the actual app side.
Added a separate mode property and checks for those.

There still seems to be some z order(?) type of issues with the
selection handles, but that's already a lot less severe.