[sailfish-browser] Avoid history button dimming while dragging. JB#63270
The OverlayListItem switches opacity to Theme.opacityLow if 'enabled' is
false, which happened here when dragging just slightly.

Was introduced 2015 in commit 47f7a6bfc0f3d334 without bug reference
or rationale. At least in current setup I don't think it should be
possible to tap the item when it shouldn't react.