Avoid DEFINE for a random base64 formatting string
No point in adding garbage to the compiler invocation when there's no
need to override this.

In addition the extra indirection was hiding that png format was
hard-coded as mime type. Fixed the explicit jpeg case but fetching
data from the network needs to still figure out the real type.